refactor(rippling): move script body to composable, CSS to its own file
Final big move to get RipplingExplorer.vue down to 287 lines.

- The entire imperative onMounted body (~2100 lines: map setup, slider
  wiring, fetch/draw functions for isochrones, freegler dots, group
  polygons, swingometer, the full ripple animation pipeline, the
  inbound digest-preview flow, URL sync, search box, the lot) moves
  into composables/rippling/setupRipplingExplorer.js as an
  `async function setupRipplingExplorer({ props, digestModal, legendMode })`
  that returns a cleanup callback.
- The `<style>` block (~435 lines) moves to RipplingExplorer.css and is
  imported from the script.

RipplingExplorer.vue is now a thin shell: template, the two refs the
composable needs to mutate (digestModal, legendMode), defineProps, and a
4-line onMounted/onUnmounted pair that calls the composable.

Sub-components and helpers from the previous three refactor commits are
also imported here.  Zero behaviour change; HMR clean.

Future commits can split setupRipplingExplorer.js further (per-feature
composables: outbound, inbound, animation, swingometer, freeglers,
groups) without touching the .vue shell again.
